(fix): issue #70- guarded so that the button only does its function if you type your username in (so it doesn't push you to root screen anymore). there's still an issue with networking/database i think because you still can't actually delete your account

+++ Resell/Views/Settings/SettingsView.swift
@@ -175,6 +175,7 @@ struct SettingsView: View {
                 }
 
             Button {
+                guard viewModel.confirmUsernameText == CurrentUserProfileManager.shared.username else { return }
                 viewModel.deleteAccount()
                 mainViewModel.selection = 0
                 mainViewModel.userDidLogin = false
@@ -187,6 +188,7 @@ struct SettingsView: View {
                     .background(Constants.Colors.errorRed)
                     .clipShape(.capsule)
             }
+            
 
             Button {
                 viewModel.togglePopup(isPresenting: false)
